remove dispatcher rpc grpc server, use in process dispather with ctx scope
