+++ endpoint_grpc.go
@@ -6,7 +6,6 @@ import (
 	plugin "github.com/hashicorp/go-plugin"
 	"github.com/treethought/roc/proto"
 	"golang.org/x/net/context"
-	"google.golang.org/grpc"
 )
 
 // EndpointGRPC is an implementation of KV that talks over RPC.
@@ -85,30 +84,9 @@ func protoToContext(p *proto.RequestContext) *RequestContext {
 
 }
 
-// TODO serve dispatcher server from kernel
-func (m *EndpointGRPC) setDispatchServer(ctx *proto.RequestContext, dispatcher Dispatcher) (stop func()) {
-	dispatchServer := &DispatcherGRPCServer{Impl: dispatcher}
-
-	var s *grpc.Server
-
-	serverFunc := func(opts []grpc.ServerOption) *grpc.Server {
-		s = grpc.NewServer(opts...)
-		proto.RegisterDispatcherServer(s, dispatchServer)
-		return s
-	}
-
-	log.Debug("starting ephemeral dispatch server for endpoiny grpc call")
-	brokerID := m.broker.NextId()
-	go m.broker.AcceptAndServe(brokerID, serverFunc)
-
-	ctx.DispatcherServer = brokerID
-	return s.Stop
-}
-
 func (m *EndpointGRPC) Evaluate(ctx *RequestContext) Representation {
 	protoCtx := newProtoContext(ctx)
 
-	m.setDispatchServer(protoCtx, ctx.Dispatcher)
 	resp, err := m.client.Evaluate(context.Background(), protoCtx)
 	if err != nil {
 		panic(err)
@@ -120,9 +98,6 @@ func (m *EndpointGRPC) Evaluate(ctx *RequestContext) Representation {
 func (m *EndpointGRPC) Source(ctx *RequestContext) Representation {
 	log.Debug("making endpoint Source grpc call", "identifier", ctx.Request.Identifier)
 	protoCtx := newProtoContext(ctx)
-	//TODO stop server
-	_ = m.setDispatchServer(protoCtx, ctx.Dispatcher)
-	// defer stop()
 
 	resp, err := m.client.Source(context.Background(), protoCtx)
 	if err != nil {
@@ -136,8 +111,6 @@ func (m *EndpointGRPC) Source(ctx *RequestContext) Representation {
 
 func (m *EndpointGRPC) Sink(ctx *RequestContext) {
 	protoCtx := newProtoContext(ctx)
-	stop := m.setDispatchServer(protoCtx, ctx.Dispatcher)
-	defer stop()
 
 	_, err := m.client.Sink(context.Background(), protoCtx)
 	if err != nil {
@@ -149,8 +122,6 @@ func (m *EndpointGRPC) Sink(ctx *RequestContext) {
 
 func (m *EndpointGRPC) New(ctx *RequestContext) Identifier {
 	protoCtx := newProtoContext(ctx)
-	stop := m.setDispatchServer(protoCtx, ctx.Dispatcher)
-	defer stop()
 
 	resp, err := m.client.New(context.Background(), protoCtx)
 	if err != nil {
@@ -162,8 +133,6 @@ func (m *EndpointGRPC) New(ctx *RequestContext) Identifier {
 
 func (m *EndpointGRPC) Delete(ctx *RequestContext) bool {
 	protoCtx := newProtoContext(ctx)
-	stop := m.setDispatchServer(protoCtx, ctx.Dispatcher)
-	defer stop()
 
 	resp, err := m.client.Delete(context.Background(), protoCtx)
 	if err != nil {
@@ -175,8 +144,6 @@ func (m *EndpointGRPC) Delete(ctx *RequestContext) bool {
 
 func (m *EndpointGRPC) Exists(ctx *RequestContext) bool {
 	protoCtx := newProtoContext(ctx)
-	stop := m.setDispatchServer(protoCtx, ctx.Dispatcher)
-	defer stop()
 
 	resp, err := m.client.Exists(context.Background(), protoCtx)
 	if err != nil {
@@ -195,24 +162,25 @@ type EndpointGRPCServer struct {
 	broker *plugin.GRPCBroker
 }
 
-func (m *EndpointGRPCServer) setDispatchClient(ctx *RequestContext, brokerID uint32) (conn *grpc.ClientConn, err error) {
-	log.Debug("setting dispatch client to handle grpc server request", "brokerID", brokerID)
-	conn, err = m.broker.Dial(brokerID)
-	if err != nil {
-		log.Error("failed to create dispatcher client conn", "error", err)
-		return nil, err
-	}
+// func (m *EndpointGRPCServer) setDispatchClient(ctx *RequestContext, brokerID uint32) (conn *grpc.ClientConn, err error) {
+// 	log.Debug("setting dispatch client to handle grpc server request", "brokerID", brokerID)
+// 	conn, err = m.broker.Dial(brokerID)
+// 	if err != nil {
+// 		log.Error("failed to create dispatcher client conn", "error", err)
+// 		return nil, err
+// 	}
+//     // defer conn.Close()
 
-	d := &DispatcherGRPC{
-		// TODO: dont use same broker?
-		// broker: &plugin.GRPCBroker{},
-		client: proto.NewDispatcherClient(conn),
-	}
-	ctx.Dispatcher = d
-	log.Debug("set context dispatcher", "dispatcher", ctx.Dispatcher)
-	return conn, nil
+// 	d := &DispatcherGRPC{
+// 		// TODO: dont use same broker?
+// 		// broker: &plugin.GRPCBroker{},
+// 		client: proto.NewDispatcherClient(conn),
+// 	}
+// 	ctx.Dispatcher = d
+// 	log.Debug("set context dispatcher", "dispatcher", ctx.Dispatcher)
+// 	return conn, nil
 
-}
+// }
 
 func (m *EndpointGRPCServer) Evaluate(ctx context.Context, req *proto.RequestContext) (*proto.Representation, error) {
 
@@ -226,12 +194,8 @@ func (m *EndpointGRPCServer) Source(ctx context.Context, req *proto.RequestConte
 	log.Debug("begining endpoint grpc source server implementation")
 
 	rocCtx := protoToContext(req)
-	conn, err := m.setDispatchClient(rocCtx, req.DispatcherServer)
-	if err != nil {
-		log.Error("error setting dispatch client", "err", err)
-		return nil, err
-	}
-	defer conn.Close()
+	// _, err := m.setDispatchClient(rocCtx, req.DispatcherServer)
+	// defer conn.Close()
 
 	rep := m.Impl.Source(rocCtx)
 
@@ -241,11 +205,6 @@ func (m *EndpointGRPCServer) Source(ctx context.Context, req *proto.RequestConte
 
 func (m *EndpointGRPCServer) Sink(ctx context.Context, req *proto.RequestContext) (*proto.Empty, error) {
 	rocCtx := protoToContext(req)
-	conn, err := m.setDispatchClient(rocCtx, req.DispatcherServer)
-	if err != nil {
-		return nil, err
-	}
-	defer conn.Close()
 
 	m.Impl.Sink(rocCtx)
 	return &proto.Empty{}, nil
@@ -253,11 +212,6 @@ func (m *EndpointGRPCServer) Sink(ctx context.Context, req *proto.RequestContext
 
 func (m *EndpointGRPCServer) New(ctx context.Context, req *proto.RequestContext) (*proto.IdentifierResponse, error) {
 	rocCtx := protoToContext(req)
-	conn, err := m.setDispatchClient(rocCtx, req.DispatcherServer)
-	if err != nil {
-		return nil, err
-	}
-	defer conn.Close()
 
 	ident := m.Impl.New(rocCtx)
 	resp := &proto.IdentifierResponse{
@@ -268,11 +222,6 @@ func (m *EndpointGRPCServer) New(ctx context.Context, req *proto.RequestContext)
 
 func (m *EndpointGRPCServer) Delete(ctx context.Context, req *proto.RequestContext) (*proto.BoolResponse, error) {
 	rocCtx := protoToContext(req)
-	conn, err := m.setDispatchClient(rocCtx, req.DispatcherServer)
-	if err != nil {
-		return nil, err
-	}
-	defer conn.Close()
 
 	result := m.Impl.Delete(rocCtx)
 	resp := &proto.BoolResponse{
@@ -283,11 +232,6 @@ func (m *EndpointGRPCServer) Delete(ctx context.Context, req *proto.RequestConte
 
 func (m *EndpointGRPCServer) Exists(ctx context.Context, req *proto.RequestContext) (*proto.BoolResponse, error) {
 	rocCtx := protoToContext(req)
-	conn, err := m.setDispatchClient(rocCtx, req.DispatcherServer)
-	if err != nil {
-		return nil, err
-	}
-	defer conn.Close()
 
 	result := m.Impl.Exists(rocCtx)
 	resp := &proto.BoolResponse{
